+++ CHANGES.md
@@ -1,6 +1,10 @@
 # Version TBD
-- Fix expire to formatTz in create account session [#4985](https://github.com/appwrite/appwrite/pull/4985)
+## Features
+- Update APIs to check X-Appwrite-Timestamp header [#5024](https://github.com/appwrite/appwrite/pull/5024)
+
+## Bugs
 - Fix not storing function's response on response codes 5xx [#4610](https://github.com/appwrite/appwrite/pull/4610)
+- Fix expire to formatTz in create account session [#4985](https://github.com/appwrite/appwrite/pull/4985)
 
 # Version 1.2.1
 ## Changes